Revert "refactor: fetch segments from API with caching (#292)"
This reverts commit 63909c18ab40cd57fe46d4e16050bd7c2ec1b2c9.
