+++ packages/backend/utils_test/CheckIfModifiedSince_test.go
@@ -1,8 +1,10 @@
-package caching
+package utils_test
 
 import (
 	"testing"
 	"time"
+
+	"github.com/FreiFahren/backend/utils"
 )
 
 func TestCheckIfModifiedSince(t *testing.T) {
@@ -47,7 +49,7 @@ func TestCheckIfModifiedSince(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			modified, err := CheckIfModifiedSince(tt.ifModifiedSince, tt.lastModified)
+			modified, err := utils.CheckIfModifiedSince(tt.ifModifiedSince, tt.lastModified)
 			if (err != nil) != tt.expectError {
 				t.Errorf("CheckIfModifiedSince() error = %v, expectError %v", err, tt.expectError)
 			}
